tested loader with -nostartfiles. Also, changed mem_addr, reg_word, and mem_word to be int32_t.

+++ src/sst/elements/mips_4kc/reg.h
@@ -25,7 +25,8 @@
 #ifndef _REG_H
 #define _REG_H
 
-typedef long reg_word;
+//typedef long reg_word;
+typedef int32_t reg_word;
